requirements.txt
Django==5.2.7
pymysql==1.0.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.6.0
python-decouple==3.8
mysqlclient==2.2.6
dj-database-url==2.1.0
